requirements.txt
openai==0.27.4
spacy==3.5.2
tiktoken==0.3.1
requests==2.28.2
Pillow==10.0.0
pyobjc==9.2
pytesseract==0.3.10